LINE-Bot/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.6
Flask-Cors==3.0.6
line-bot-sdk